requirements.txt
asgiref==3.8.1
Brotli==1.1.0
click==8.1.7
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
gunicorn==22.0.0
h11==0.14.0
inflection==0.5.1
packaging==24.0
pillow==10.3.0
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
rest-framework-simplejwt==0.0.2
sqlparse==0.5.0
uritemplate==4.1.1
uvicorn==0.29.0
whitenoise==6.6.0
